refactor: friend quack v2 마이그레이션
